Refactor tooltip to provide custom and regexp tooltip
