+++ src/index.ts
@@ -26,7 +26,7 @@ export {
 export { default as SingleConnectedAutocompleteField } from './InputField/Select/Autocomplete/Connected/Single';
 export { default as MultiConnectedAutocompleteField } from './InputField/Select/Autocomplete/Connected/Multi';
 export { default as SearchField } from './InputField/Search';
-export { default as PersistentTooltip } from './InputField/Search/PersistentTooltip';
+export { default as RegexpHelpTooltip } from './InputField/Search/RegexpHelpTooltip';
 export { default as TextField } from './InputField/Text';
 export { default as SelectField, SelectEntry } from './InputField/Select';
